requirements.txt
Flask == 1.1.2
gunicorn == 19.9.0
Werkzeug == 1.0.1
MarkupSafe >= 0.23
Jinja2 == 2.11.1
tensorflow-cpu == 2.5.0
numpy == 1.18.1
pillow == 7.1.2
